+++ R/provider-azure.R
@@ -149,7 +149,7 @@ method(chat_request, ProviderAzure) <- function(provider,
   if (nchar(provider@api_key)) {
     req <- req_headers_redacted(req, `api-key` = provider@api_key)
   }
-  req <- req_headers(req, !!!provider@credentials(), .redact = "Authorization")
+  req <- ellmer_req_credentials(req, provider@credentials)
   req <- req_retry(req, max_tries = 2)
   req <- ellmer_req_timeout(req, stream)
   req <- req_error(req, body = function(resp) {