+++ tests/testthat/test-provider-gemini-upload.R
@@ -1,0 +1,8 @@
+test_that("can upload a file an ask questions about it", {
+  upload <- gemini_upload(test_path("apples.pdf"))
+
+  chat <- chat_gemini()
+  response <- chat$chat("What's the title of this document?", upload)
+  expect_match(response, "Apples are tasty")
+  expect_match(chat$chat("What apple is not tasty?"), "red delicious")
+})